refactor : 수정사항 테스트 코드 반영
